requirements.txt
flask==2.3.3
edge-tts==7.2.3
pandas>=2.2.0
openpyxl==3.1.2
pyngrok==7.0.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
